feat(chart): Set `allowPrivilegeEscalation: false`
The container ports needed to be changed to 8080 and 8443 for this purpose. Service ports remain unchanged.
The capability NET_BIND_SERVICE is still needed. Otherwise the nginx process can not even be started due to the file capability on `/usr/sbin/nginx` in the image.
There is no change in the image.